requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atlassian-python-api==3.41.21
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
Deprecated==1.2.18
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==5.3.1
Markdown==3.7
markdown-to-confluence==0.3.1
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.4.1
mcp-atlassian==0.2.4
oauthlib==3.2.2
outcome==1.3.0.post0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pymdown-extensions==10.14.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.46.1
trio==0.29.0
types-html5lib==1.1.11.20241018
types-lxml==2025.3.4
types-Markdown==3.7.0.20241204
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
